+++ src/rer/solrpush/solr.py
@@ -6,6 +6,7 @@ from plone.indexer.interfaces import IIndexableObject
 from plone.registry.interfaces import IRegistry
 from pysolr import SolrError
 from rer.solrpush import _
+from rer.solrpush.interfaces.adapter import IExtractFileFromTika
 from rer.solrpush.interfaces.settings import IRerSolrpushSettings
 from zope.component import getUtility
 from zope.component import queryMultiAdapter
@@ -198,6 +199,17 @@ def extract_fields(nodes):
     return fields
 
 
+def attachment_to_index(item):
+    """
+    If item has a provider to extract text, return the file to be indexed
+    """
+    try:
+        provider = IExtractFileFromTika(item)
+        return provider.get_file_to_index()
+    except TypeError:
+        return None
+
+
 def is_solr_active():
     """Just checking if solr indexing is set to active in control panel"""
     return get_setting(field="active")
@@ -224,7 +236,6 @@ def create_index_dict(item):
     """Restituisce un dizionario pronto per essere 'mandato' a SOLR per
     l'indicizzazione.
     """
-
     index_fields = get_index_fields(field="index_fields")
     frontend_url = get_setting(field="frontend_url")
 
@@ -265,6 +276,7 @@ def create_index_dict(item):
         )
     else:
         index_me["url"] = item.absolute_url()
+    index_me["attachment"] = attachment_to_index(item)
     return index_me
 
 
@@ -307,6 +319,7 @@ def generate_query(
         "json.nl": "arrmap",
     }
     solr_query.update(extract_from_query(query=query))
+
     if not solr_query["q"]:
         solr_query["q"] = "*:*"
     if filtered_sites:
@@ -415,11 +428,41 @@ def push_to_solr(item_or_obj):
             item_or_obj = create_index_dict(item_or_obj)
         else:
             item_or_obj = None
-    if item_or_obj:
-        solr.add([item_or_obj])
-        return True
-    else:
+    if not item_or_obj:
         return False
+    attachment = item_or_obj.pop("attachment", None)
+    if attachment:
+        add_with_attachment(
+            solr=solr, attachment=attachment, fields=item_or_obj
+        )
+    else:
+        solr.add([item_or_obj])
+    return True
+
+
+def add_with_attachment(solr, attachment, fields):
+    params = {
+        "extractOnly": "false",
+        "lowernames": "false",
+        "fmap.content": "SearchableText",
+        "fmap.title": "SearchableText",
+        "fmap.created": "ignore_created",
+        "fmap.modified": "ignore_modified",
+        "literalsOverride": "false",
+        "commit": "true",
+    }
+    params.update(
+        {
+            "literal.{key}".format(key=key): value
+            for (key, value) in fields.items()
+        }
+    )
+    return solr._send_request(
+        "post",
+        "update/extract",
+        body=params,
+        files={"file": ("extracted", attachment)},
+    )
 
 
 def remove_from_solr(uid):